# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
			"AreaTypeSimple": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					}
				},
				"required": [
					"identifier",
					"label",
					"isGroup"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaArray": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"areaType": {
						"$ref": "#/components/schemas/AreaTypeSimple"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"areaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaGetResponse": {
				"properties": {
					"area-array": {
						"items": {
							"$ref": "#/components/schemas/AreaArray"
						},
						"type": "array"
					}
				},
				"required": [
					"area-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaTypeArray": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"isGroup"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaTypeIdentifer": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					},
					"areaType": {
						"$ref": "#/components/schemas/AreaTypeArray"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"isGroup",
					"areaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BoundingBox": {
				"properties": {
					"minLatitude": {
						"type": "number",
						"format": "double"
					},
					"maxLatitude": {
						"type": "number",
						"format": "double"
					},
					"minLongitude": {
						"type": "number",
						"format": "double"
					},
					"maxLongitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"minLatitude",
					"maxLatitude",
					"minLongitude",
					"maxLongitude"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Type": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"identifier",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Area": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"areaType": {
						"$ref": "#/components/schemas/AreaTypeIdentifer"
					},
					"boundingBox": {
						"$ref": "#/components/schemas/BoundingBox"
					},
					"governedBy": {
						"$ref": "#/components/schemas/GovernedBy"
					},
					"memberOf": {
						"items": {
							"$ref": "#/components/schemas/DefaultComparisonGroup"
						},
						"type": "array"
					},
					"ownerOf": {
						"items": {
							"$ref": "#/components/schemas/DefaultComparisonGroup"
						},
						"type": "array"
					},
					"defaultComparisonGroup": {
						"$ref": "#/components/schemas/DefaultComparisonGroup"
					},
					"childOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeIdentifer"
						},
						"type": "array"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"areaType",
					"boundingBox",
					"governedBy",
					"memberOf",
					"ownerOf",
					"defaultComparisonGroup",
					"childOf"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaIdentifierGetResponse": {
				"properties": {
					"area": {
						"$ref": "#/components/schemas/Area"
					}
				},
				"required": [
					"area"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaIdentifierBoundGetResponse": {
				"properties": {
					"boundingBox": {
						"$ref": "#/components/schemas/BoundingBox"
					}
				},
				"required": [
					"boundingBox"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaTypeGetResponse": {
				"properties": {
					"areaType-array": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					}
				},
				"required": [
					"areaType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaTypeCurrentGetResponse": {
				"properties": {
					"areaType-array": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					}
				},
				"required": [
					"areaType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaTypeFilterDataGetResponse": {
				"properties": {
					"areaType-array": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					}
				},
				"required": [
					"areaType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaTypeArrayVerbose": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					},
					"isMain": {
						"type": "boolean"
					},
					"childOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"parentOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"memberOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"sameAs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"isGroup",
					"isMain",
					"childOf",
					"parentOf",
					"memberOf",
					"sameAs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaTypeVerboseGetResponse": {
				"properties": {
					"areaType-array": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArrayVerbose"
						},
						"type": "array"
					}
				},
				"required": [
					"areaType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AreaType": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"shortLabel": {
						"type": "string"
					},
					"longLabel": {
						"type": "string"
					},
					"isGroup": {
						"type": "boolean"
					},
					"isMain": {
						"type": "boolean"
					},
					"childOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"parentOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"memberOf": {
						"items": {
							"$ref": "#/components/schemas/AreaTypeArray"
						},
						"type": "array"
					},
					"sameAs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"identifier",
					"label",
					"shortLabel",
					"longLabel",
					"isGroup",
					"isMain",
					"childOf",
					"parentOf",
					"memberOf",
					"sameAs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAreaTypeIdentifierGetResponse": {
				"properties": {
					"areaType": {
						"$ref": "#/components/schemas/AreaType"
					}
				},
				"required": [
					"areaType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ComparisonGroup": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"isCustom": {
						"type": "boolean"
					},
					"label": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					}
				},
				"required": [
					"identifier",
					"isCustom",
					"label",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupsGetResponse": {
				"properties": {
					"comparisonGroup-array": {
						"items": {
							"$ref": "#/components/schemas/ComparisonGroup"
						},
						"type": "array"
					}
				},
				"required": [
					"comparisonGroup-array"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ComparisonGroupIdentifier": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"isCustom": {
						"type": "boolean"
					},
					"label": {
						"type": "string"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"owners": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"isMain": {
						"type": "boolean"
					},
					"isHidden": {
						"type": "boolean"
					},
					"bottomTier": {
						"$ref": "#/components/schemas/Tier"
					},
					"topTier": {
						"$ref": "#/components/schemas/Tier"
					},
					"type": {
						"$ref": "#/components/schemas/Type"
					}
				},
				"required": [
					"identifier",
					"isCustom",
					"label",
					"members",
					"owners",
					"isMain",
					"isHidden",
					"bottomTier",
					"topTier",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupsIdentifierGetResponse": {
				"properties": {
					"comparisonGroup": {
						"$ref": "#/components/schemas/ComparisonGroupIdentifier"
					}
				},
				"required": [
					"comparisonGroup"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ComparisonGroupTypeArray": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"identifier",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupTypesGetResponse": {
				"properties": {
					"comparisonGroupType-array": {
						"items": {
							"$ref": "#/components/schemas/ComparisonGroupTypeArray"
						},
						"type": "array"
					}
				},
				"required": [
					"comparisonGroupType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupTypesVerboseGetResponse": {
				"properties": {
					"comparisonGroupType-array": {
						"items": {
							"$ref": "#/components/schemas/ComparisonGroupTypeArray"
						},
						"type": "array"
					}
				},
				"required": [
					"comparisonGroupType-array"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ComparisonGroupType": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"identifier",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IComparisonGroupTypesIdentifierGetResponse": {
				"properties": {
					"comparisonGroupType": {
						"$ref": "#/components/schemas/ComparisonGroupType"
					}
				},
				"required": [
					"comparisonGroupType"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "query",
				"name": "ApplicationKey"
			}
		}
	},
	"info": {
		"title": "Local Government Association API",
		"version": "v1",
		"description": "Access to 1 billion+ metric values for English areas drawn from approximately 50 organisations publishing metrics that describe local authorities and their component areas. (DkxYeiqbmsGVMHOkfbbUGUEomQYkfaiiEJCBDEtC)",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/areas": {
			"get": {
				"operationId": "GetArea",
				"responses": {
					"200": {
						"description": "IAreaGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaGetResponse"
								}
							}
						}
					}
				},
				"description": "Due to the large number of areas an empty call returns no results. Results are limited to first 500 - subsequent results can be found using the \"pageNumber\" attribute",
				"summary": "Geographical areas used by LG Inform.",
				"tags": [
					"area"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Filter areas by comparison group and metric type's applicable area types",
						"in": "query",
						"name": "applicableToMetricType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by area type and parent area",
						"in": "query",
						"name": "area",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "All areas of a particular area type",
						"in": "query",
						"name": "areaType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Show areas for a spceified latitude and longitude (alongside lon parameter)",
						"in": "query",
						"name": "lat",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Show areas for a spceified latitude and longitude (alongside lat parameter)",
						"in": "query",
						"name": "lon",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If retrieving more than 20,000 areas or 500 shapes this can be used to return subsequent pages",
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Number if items to be returned in each call up to a limit of 500, which is the default",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The search text can be either a postcode, a postcode district (the first half of the postcode), or a text string. Defaults to local authorities whether the areaType parameter is not specified",
						"in": "query",
						"name": "searchText",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/areas/{identifier}": {
			"get": {
				"operationId": "GetAreaIdentifier",
				"responses": {
					"200": {
						"description": "IAreaIdentifierGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaIdentifierGetResponse"
								}
							}
						}
					}
				},
				"summary": "The identifier is a GSS code, or comparison group identifier, used to find details about a particular geographical area",
				"tags": [
					"area"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(E09000010)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/areas/{identifier}/bounds": {
			"get": {
				"operationId": "GetAreaIdentifierBounds",
				"responses": {
					"200": {
						"description": "IAreaIdentifierBoundGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaIdentifierBoundGetResponse"
								}
							}
						}
					}
				},
				"summary": "Co-ordinates for the minimum and maxim longitude and latitudes for a particular area or group of areas",
				"tags": [
					"area"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(E09000006)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/areaTypes": {
			"get": {
				"operationId": "GetAreaTypes",
				"responses": {
					"200": {
						"description": "IAreaTypeGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaTypeGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns area types, including detailed area types such as MBC and London borough.",
				"tags": [
					"areaTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Limit to the area types within a particular area (E09000006)",
						"in": "query",
						"name": "area",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/areaTypes/current": {
			"get": {
				"operationId": "GetAreaTypesCurrent",
				"responses": {
					"200": {
						"description": "IAreaTypeCurrentGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaTypeCurrentGetResponse"
								}
							}
						}
					}
				},
				"summary": "Current user's area type. Requires a call using OAuth to retrieve data",
				"tags": [
					"areaTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": []
			}
		},
		"/v1/areaTypes/filter-data": {
			"get": {
				"operationId": "GetAreaTypesFilterData",
				"responses": {
					"200": {
						"description": "IAreaTypeFilterDataGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaTypeFilterDataGetResponse"
								}
							}
						}
					}
				},
				"summary": "Area types which have data for a particular metric type and time period",
				"tags": [
					"areaTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Show which area levels have data for a particular metric type (3214)",
						"in": "query",
						"name": "metricType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Show which area levels have data for a particular time period (cal_2005)",
						"in": "query",
						"name": "period",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/areaTypes/verbose": {
			"get": {
				"operationId": "GetAreaTypesVerbose",
				"responses": {
					"200": {
						"description": "IAreaTypeVerboseGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaTypeVerboseGetResponse"
								}
							}
						}
					}
				},
				"summary": "List of all details for all area types",
				"tags": [
					"areaTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Filter down to main official geographies",
						"in": "query",
						"name": "isMain",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/areaTypes/{identifier}": {
			"get": {
				"operationId": "GetAreaTypesWithIdentifier",
				"responses": {
					"200": {
						"description": "IAreaTypeIdentifierGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAreaTypeIdentifierGetResponse"
								}
							}
						}
					}
				},
				"summary": "Details about a particular area type",
				"tags": [
					"areaTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(District)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/comparisonGroups": {
			"get": {
				"operationId": "GetComparisonGroup",
				"responses": {
					"200": {
						"description": "IComparisonGroupsGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupsGetResponse"
								}
							}
						}
					}
				},
				"summary": "Comparison groups are area groupings of different types, such as all district local authorities, or all fire organisations in England. Without parameters, all groups, excluding custom groups, are returned",
				"tags": [
					"comparisonGroups"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Only the main comparison groups, so includes groupings by authority type, but excludes things like groupings by near neighbours",
						"in": "query",
						"name": "mainOnly",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Groups of which the area is a member, but not necessarily an owner. If area 1 is a near neighbour of area 2, area 1 would be a member of the near neighbour group (along with the other near neighbours) and area 2 would be the owner of the group",
						"in": "query",
						"name": "memberArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Group which the area owns, but is not necessairly a member. An area is generally the owner of a group if the group is a main group, such as all local authorities, or is the area to which a group relates, such as near neighbours of an area",
						"in": "query",
						"name": "ownerArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Comparison groups of a specified comparison group type for a specified area. For example, which Experian Mosaic near neighbours a particular area has",
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/comparisonGroups/{identifier}": {
			"get": {
				"operationId": "GetComparisonGroupIdentifier",
				"responses": {
					"200": {
						"description": "IComparisonGroupsIdentifierGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupsIdentifierGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns the details for a particular comparison group",
				"tags": [
					"comparisonGroups"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(AllLAInRegion_London)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/comparisonGroupTypes": {
			"get": {
				"operationId": "GetComparisonGroupTypes",
				"responses": {
					"200": {
						"description": "IComparisonGroupTypesGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupTypesGetResponse"
								}
							}
						}
					}
				},
				"summary": "Show types of comparison group available",
				"tags": [
					"comparisonGroupTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "Filter groups by area membership",
						"in": "query",
						"name": "memberArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter groups by a specific area owner",
						"in": "query",
						"name": "ownerArea",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/comparisonGroupTypes/verbose": {
			"get": {
				"operationId": "GetComparisonGroupTypesVerbose",
				"responses": {
					"200": {
						"description": "IComparisonGroupTypesVerboseGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupTypesVerboseGetResponse"
								}
							}
						}
					}
				},
				"summary": "Show full details of comparison group types available",
				"tags": [
					"comparisonGroupTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": []
			}
		},
		"/v1/comparisonGroupTypes/{identifier}": {
			"get": {
				"operationId": "GetComparisonGroupTypesIdentifier",
				"responses": {
					"200": {
						"description": "IComparisonGroupTypesIdentifierGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComparisonGroupTypesIdentifierGetResponse"
								}
							}
						}
					}
				},
				"summary": "Show details of a specified comparison group",
				"tags": [
					"comparisonGroupTypes"
				],
				"security": [
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"description": "(AllLaInRegion)",
						"in": "path",
						"name": "identifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}